Keep context, and single-render paused video.
Previously the gl context was dropped when the video was paused, causing the rendering to disappear. Now, this only happens on disconnect, and when the context is out of the viewport. Additionally, when a paused video enters the viewport, it renders a single paused frame with the new gl context.

Fixes #5.